Fixes #20, switches to git source for nodejs ckbk.
Previously, the version of NPM being pulled down by the cookbook was broken on CentOS. The git repo was updated, but no release was made. This commit switches the Berksfile to source the cookbook from git instead, causing a version that isn't broken on CentOS to be installed. Having a working NPM is a critical path item.